% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\name{eq_clean_data}
\alias{eq_clean_data}
\title{Data Cleaning}
\usage{
eq_clean_data(raw_data)
}
\arguments{
\item{raw_data}{Raw NOAA data frame}
}
\value{
This function returns a tibble containing the cleaned earthquake data:
\itemize{
  \item A DATE column created by uniting the year, month, day and converting
        it to the Date class
  \item LATITUDE and LONGITUDE columns converted to numeric class
  \item LOCATION_NAME_cleaned column from LOCATION_NAME by stripping out the
        country name and in title case
}
}
\description{
This function cleans the NOAA Significant Earthquake Database
from \url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
\examples{
\dontrun{
noaa <- readr::read_delim(system.file("extdata", "signif.txt",
        package = "NOAAnalysis"), delim = "\\t") \%>\% eq_clean_data()
}

}
\references{
For more information on NOAA data see
\url{https://www.ngdc.noaa.gov/nndc/struts/form?t=101650&s=1&d=1}
}
