% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\name{eq_create_label}
\alias{eq_create_label}
\title{HTML Labeling}
\usage{
eq_create_label(data)
}
\arguments{
\item{data}{The filtered data frame with earthquakes to visualize}
}
\value{
This function returns a HTML label
}
\description{
Creates an HTML label that can be used as the annotation text
in the leaflet map
}
\details{
This function puts together a character string for each earthquake
that showes the cleaned location, the magnitude, and the total number of deaths
(if not missing).
}
\examples{
\dontrun{
readr::read_delim(system.file("extdata", "signif.txt",
       package = "NOAAnalysis"), delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & YEAR >= 2000) \%>\%
dplyr::mutate(popup_text = eq_create_label(.)) \%>\%
eq_map(annot_col = "popup_text")
}

}
