% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\name{eq_location_clean}
\alias{eq_location_clean}
\title{String Cleaning (Helper Function)}
\usage{
eq_location_clean(raw_string)
}
\arguments{
\item{raw_string}{Raw string of LOCATION_NAME}
}
\value{
This function returns a clean string of LOCATION_NAME
}
\description{
This function cleans the LOCATION_NAME string by stripping out
the country name and converts names to title case
}
\examples{
\dontrun{eq_location_clean("SYRIA: UGARIT")}

}
