% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\name{eq_map}
\alias{eq_map}
\title{Visualize the Earthquakes in Space}
\usage{
eq_map(data, annot_col)
}
\arguments{
\item{data}{The filtered data frame with earthquakes to visualize}

\item{annot_col}{Column containing annotation data}
}
\value{
This function returns a HTML widget object
}
\description{
Maps and annotates the epicenters
}
\details{
Each earthquake is shown with a circle, and the radius of the circle
is proportional to the earthquake's magnitude. Each point is annotated in pop
up window containing annotation data stored in the \code{annot_col} column
of \code{data}.
}
\examples{
\dontrun{
readr::read_delim(system.file("extdata", "signif.txt",
       package = "NOAAnalysis"), delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "MEXICO" & YEAR >= 2000) \%>\%
eq_map(annot_col = "DATE")
}

}
