% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\docType{data}
\name{GeomTimeline}
\alias{GeomTimeline}
\title{GeomTimeline Class Construction}
\description{
Construct the new class corresponding to the \code{geom_timeline} geom
}
\note{
Not exported to the user
}
\keyword{datasets}
