% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\docType{data}
\name{GeomTimelineLabel}
\alias{GeomTimelineLabel}
\title{GeomTimelineLabel Class Construction}
\description{
Construct the new class corresponding to the
\code{geom_timeline_label} geom
}
\note{
Not exported to the user
}
\keyword{datasets}
