% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\docType{data}
\name{StatTimeline}
\alias{StatTimeline}
\title{Date Filtering Stat Construction}
\description{
This stat corresponds to the \code{geom_timeline} and
\code{geom_timeline_label} geoms and filters the earthquakes ranging
from \code{xmin} to \code{xmax} dates
}
\note{
Not exported to the user
}
\keyword{datasets}
