% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NOAA_functions.R
\docType{data}
\name{theme_timeline}
\alias{theme_timeline}
\title{Theme for geom_timeline Plot}
\value{
Theme object
}
\description{
Simple theme for \code{geom_timeline} plot
}
\examples{
\dontrun{
readr::read_delim(system.file("extdata", "signif.txt",
       package = "NOAAnalysis"), delim = "\\t") \%>\%
eq_clean_data() \%>\%
dplyr::filter(COUNTRY == "USA" & YEAR >= 2000) \%>\%
ggplot2::ggplot(ggplot2::aes(x = DATE, y = COUNTRY,
         xmin = as.Date("2000-01-01"), xmax = as.Date("2016-01-01"),
         size = as.numeric(EQ_PRIMARY), color = as.numeric(TOTAL_DEATHS))) +
geom_timeline() +
theme_timeline +
ggplot2::labs(colour = "# deaths", size = "Richter scale value")
}

}
\keyword{datasets}
